[IMP] Mail: Move suggestion list code inside a model
This will reduce composer model complexity and allow to reuse suggestion code
inside other component.

task-2447140
